Removing unused code
updating video.js

Extracting out services/index
