+++ package.json
@@ -70,7 +70,7 @@
     "ramda": "^0.27.1",
     "redis": "^3.1.2",
     "uuid": "^8.3.2",
-    "video.js": "^7.11.8",
+    "video.js": "^7.12.1",
     "winston": "^3.3.3"
   },
   "devDependencies": {
